Add warning message for when projects override implicit PackageReference items
