+++ src/Tasks/Microsoft.NET.Build.Tasks/Microsoft.NET.Build.Tasks.csproj
@@ -36,7 +36,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CheckForDuplicateItems.cs" />
-    <Compile Include="CheckForDuplicateItemsToRemove.cs" />
+    <Compile Include="CheckForImplicitPackageReferenceOverrides.cs" />
     <Compile Include="DiagnosticMessageSeverity.cs" />
     <Compile Include="DiagnosticsHelper.cs" />
     <Compile Include="NETSdkError.cs" />
