+++ src/Tasks/Microsoft.NET.Build.Tasks/Resources/Strings.resx
@@ -228,4 +228,7 @@
   <data name="DuplicateItemsError" xml:space="preserve">
     <value>Duplicate '{0}' items were included. The .NET SDK includes '{0}' items from your project directory by default. You can either remove these items from your project file, or set the '{1}' property to '{2}' if you want to explicitly include them in your project file. The duplicate items were: {3}</value>
   </data>
+  <data name="PackageReferenceOverrideWarning" xml:space="preserve">
+    <value>A PackageReference for '{0}' was included in your project. This package is implicitly referenced by the .NET SDK and you do not typically need to reference it from your project. For more information, see {1}</value>
+  </data>
 </root>
